GobCollection maintains a Dictionary for efficiently finding gobs by their IDs. Also removed dead code.
This commit fixes the bug where a client couldn't join an ongoing arena if there were lots of gobs active (at least when profiling the code). The client would hang for 10+ seconds trying to find the gobs for god knows what, until the server cut the connection due to timeout. After this change, joining an arena like that is only a matter of around one second.