Remove ManagementServer & WebData to make way for Steam
I kept these before because I wasn't sure of their role in the future, but
now we are going to replace all previous "global" functionality
with steam functionality.

Removing this code will make it easier for me to see what I'm doing
when I'm integrating Steam networking and multiplayer features.