Extract abstract interfaces NetworkEngine and Connection
Moved some functionality behind the interface; specifically
connection result handling that needed specifics of the
connection.

Some abstraction related to StartServer and StartClient was
dismantled to make the NetworkEngine abstraction stronger.
Similar abstraction can be added back later.

The Spectator was not using its IPAddress much so it was
turned into a string to avoid needing to encode specifics of
the connection technology in the Spectator. I'm also not sure
what the "same spectator from different address is doing".
Possibly some kind of cheat check or something.