Move PlayerID & IsLoggedIn to Spectator
These were in SpectatorStats previously because they were related to
the management server and stats collection.

The PlayerId is now a securely hashed SteamID when in Steam mode and a
securely hashed connection address string when in the raw networking
mode.

Secure hashing is used to make sure identifying information about
players does not unnecessarily leak to other clients.